+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Changes are categorized as follows:
 
 ## [Unreleased]
 
-* 8 minor enhancements:
+* 9 minor enhancements:
 
   * docs: add github paths
   * deps: add debug
@@ -23,6 +23,7 @@ Changes are categorized as follows:
   * deps: add root-model
   * deps: add schema-migrations
   * docs: simpler setup with npm
+  * config: improve db interface
 
 ## 2018-05-03, Version 1.0.1